+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/init/TelemetryClientInitializer.java
@@ -23,6 +23,8 @@ package com.microsoft.applicationinsights.agent.internal.init;
 
 import static java.util.concurrent.TimeUnit.MINUTES;
 
+import com.azure.monitor.opentelemetry.exporter.implementation.configuration.ConnectionString;
+import com.azure.monitor.opentelemetry.exporter.implementation.configuration.StatsbeatConnectionString;
 import com.microsoft.applicationinsights.agent.internal.common.PropertyHelper;
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
 import com.microsoft.applicationinsights.agent.internal.configuration.Configuration;
@@ -116,7 +118,13 @@ public class TelemetryClientInitializer {
     String connectionString = configuration.connectionString;
 
     if (connectionString != null) {
-      telemetryClient.setConnectionString(connectionString);
+      ConnectionString connectionStringObj = ConnectionString.parse(connectionString);
+      telemetryClient.setConnectionString(connectionStringObj);
+      telemetryClient.setStatsbeatConnectionString(
+          StatsbeatConnectionString.create(
+              connectionStringObj,
+              configuration.internal.statsbeat.instrumentationKey,
+              configuration.internal.statsbeat.endpoint));
     }
   }