+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/init/RpConfigurationPollingTest.java
@@ -23,6 +23,7 @@ package com.microsoft.applicationinsights.agent.internal.init;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.azure.monitor.opentelemetry.exporter.implementation.configuration.ConnectionString;
 import com.microsoft.applicationinsights.agent.internal.configuration.Configuration;
 import com.microsoft.applicationinsights.agent.internal.configuration.RpConfiguration;
 import com.microsoft.applicationinsights.agent.internal.legacysdk.BytecodeUtilImpl;
@@ -79,8 +80,10 @@ class RpConfigurationPollingTest {
     rpConfiguration.lastModifiedTime = 0;
 
     TelemetryClient telemetryClient = TelemetryClient.createForTest();
-    telemetryClient.setConnectionString("InstrumentationKey=00000000-0000-0000-0000-000000000000");
-    AppIdSupplier appIdSupplier = new AppIdSupplier(telemetryClient);
+    ConnectionString connectionString =
+        ConnectionString.parse("InstrumentationKey=00000000-0000-0000-0000-000000000000");
+    telemetryClient.setConnectionString(connectionString);
+    AppIdSupplier appIdSupplier = new AppIdSupplier(connectionString);
 
     BytecodeUtilImpl.samplingPercentage = 100;
 
@@ -113,8 +116,10 @@ class RpConfigurationPollingTest {
     rpConfiguration.lastModifiedTime = 0;
 
     TelemetryClient telemetryClient = TelemetryClient.createForTest();
-    telemetryClient.setConnectionString("InstrumentationKey=00000000-0000-0000-0000-000000000000");
-    AppIdSupplier appIdSupplier = new AppIdSupplier(telemetryClient);
+    ConnectionString connectionString =
+        ConnectionString.parse("InstrumentationKey=00000000-0000-0000-0000-000000000000");
+    telemetryClient.setConnectionString(connectionString);
+    AppIdSupplier appIdSupplier = new AppIdSupplier(connectionString);
 
     BytecodeUtilImpl.samplingPercentage = 100;